drm: block userspace under allocating buffer and having drivers overwrite it (v2)
With the current screwed but its ABI, ioctls for the drm, Linus pointed
out that we could allow userspace to specify the allocation size, but we
pass it to the driver which then uses it blindly to store a struct. Now
if userspace specifies the allocation size as smaller than the driver
needs, the driver can possibly overwrite memory.

This patch restructures the driver ioctls so we store the structure size
we are expecting, and make sure we allocate at least that size. The copy
from/to userspace are still restricted to the size the user specifies,
this allows ioctl structs to grow on both sides of the equation.

Up until now we didn't really use the DRM_IOCTL defines in the kernel,
so this cleans them up and adds them for nouveau.

v2:
fix nouveau pushbuf arg (thanks to Ben for pointing it out)

Reported-by: Linus Torvalds <torvalds@linuxfoundation.org>
Signed-off-by: Dave Airlie <airlied@redhat.com>