kernel wastes huge amounts of memory due to CONFIG_IMA
Previously, Red Hat Enterprise Linux 6 enabled the  CONFIG_IMA option in
the kernel. This caused the kernel to track all inodes in the system in
a radix tree, leading to a huge waste of memory. With this update, an
optimized version of a tree (rbtree) is used and memory is no longer
wasted.

Red Hat fix for https://bugzilla.redhat.com/show_bug.cgi?id=644636