kernel: BUG: warning at drivers/char/tty_audit.c:55/tty_audit_buf_free()
Using the pam_tty_audit.so module (which enables or disables TTY
auditing for specified users) in the /etc/pam.d/sudo file and in the
/etc/pam.d/system-auth file when the audit package is not installed
resulted in soft lock-ups on CPUs. As a result, the kernel became
unresponsive. This was due to the kernel exiting immediately after TTY
auditing was disabled, without emptying the buffer, which caused the
kernel to spin in a loop, copying 0 bytes at each iteration and
attempting to push each time without any effect. With this update, a
locking mechanism is introduced to prevent the aforementioned behavior.

https://bugzilla.redhat.com/show_bug.cgi?id=684275